requirements.txt
streamlit
rembg
onnxruntime
pillow
numpy
onnxruntime-gpu
